Projects/1. LangChain 101 - from Zero to Hero/requirements.txt
langchain==0.0.208
deeplake
openai
tiktoken
deeplake
python-dotenv